lowmemorykiller: module to kill processes when memory gets tight
Provides more control to userspace than the standard OOM killer.

Up to six levels are supported. By default processes
with a oomkilladj value of 0 or higer are killed if
the sum if the page cache is less than 6MB when we
run out of memory, 1 or higher if it is less than 8MB,
6 or higher if it less than 16MB and 12 or higher if
it is less than 16MB.

Signed-off-by: Brian Swetland <swetland@google.com>