Fix suspend_console/resume_console use only one semaphore.
This fixes a race if another thread acquired the console
while it was suspended and released it after it was resumed
the secondary console semaphre would be left locked, and the
primary semaphore would be released twice. This in turn caused
the console switch on suspend or resume to hang forever.

Note that suspend_console does not actually lock the console
for clients that use acquire_console_sem, it only locks it for
clients that use try_acquire_console_sem. If we change
suspend_console to fully lock the console, then the kernel
may deadlock on suspend.

Signed-off-by: Brian Swetland <swetland@google.com>