Bluetooth: Pause RFCOMM tx connection when encryption is dropped.
Some headsets drop the encryption when doing the role change.
We allow a grace period for the encryption to be reestablished.
During this time we pause the RFCOMM tx traffic and if encryption
is not reestablished we drop the connection.

Based on a original change by: Marcel Holtmann <marcel@holtmann.org>

Signed-off-by: Jaikumar Ganesh <jaikumar@google.com>