wl12xx: fix potential buffer overflow in testmode nvs push
We were allocating the size of the NVS file according to the chip ID
and not checking whether the length of the buffer passed was correct
before copying it into the allocated memory.  This is a security hole
because buffer overflows can occur if the userspace passes a bigger
file than what is expected.

With this patch, we check if the size of the data passed from
userspace matches the size required by the chip.

Reported-by: Ido Yariv <ido@wizery.com>
Signed-off-by: Luciano Coelho <coelho@ti.com>